navigation-fragment | Move main files
